+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This changelog covers what's changed in the MPDX APIs.
 
+## 20 April 2017
+- Contact suggested_changes used to sometimes suggest a blank value for status, this was a bug that has been fixed.
+
 ## 18 April 2017
 - Added direct deposit to contact endpoints.
 - Updated translations and added it, th and id.